move document client actions
